+++ packages/contracts/contracts/test/unit/BalanceManager.t.sol
@@ -56,7 +56,8 @@ contract BalanceManagerTest is Test {
 
         balanceManager.initialize(
             address(teller),
-            address(subtreeUpdateVerifier)
+            address(subtreeUpdateVerifier),
+            address(0x1)
         );
 
         // NOTE: TestBalanceManager implements IHandler so we can test with